app/requirements.txt
Django==1.8
paho-mqtt==1.1